Move all SDL2_gfx low level module to a single gfx::ffi module
